ftrace: Allow setting the interval for record mode
The trace-cmd record mode allows setting the interval to determine how
often it will write to the trace file. The default is 1000 (us, ie. 1ms).

Add a collector parameter to allow setting that interval. This will only
have an effect when trace_cmd_mode is 'record' as the option does not
exist for 'start' mode.

Signed-off-by: Kajetan Puchalski <kajetan.puchalski@arm.com>