utils/asyn: Ensure the async generators inside context managers are fully consumed on a single event loop
Context managers implemented using contextlib.asynccontextmanager() rely
on an async generators. This async generator must be fully consumed on a
single event loop, otherwise:

1. The async generator will be closed after the __enter__ run() call, as
   event loops close all async generators upon closing.

2. The async generator would end up being migrated from one event loop
   to another, which is not going to work.

Sidestep both issues by ensuring that the same event loop is in use for
both __enter__() and __exit__(). If necessary, _AsyncPolymorphicCM()
creates and manages its own event loop to ensure this is the case.