utils/android: Fix AdbConnection.execute(check_exit_code=False) default
Switch to have check_exit_code=True just like any other connection.

The current behavior will not raise any exception if the command returns
a non-zero exit code. This leads to failed attempt at parsing the
output, which is now an error message rather than the expected data.
Worse, the caller may never realize the command failed. This is
especially bad as that behavior will only manifest itself when things go
wrong, which is not the majority of the time, leading to code that seems
to work ok, but does not handle failure properly (like a shell script).
Lastly, since this is at odds with all the other connection types,
generic code will typically assume check_exit_code=True by default and
end up being buggy when used in conjunction of the AdbConnection.